feat:レイヤーツリーの初期値を変更 #7
